# hot-wheels-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hot Wheels Collector API
  version: 1.0.0
  description: |
    # Hot Wheels Collector Platform API

    A comprehensive REST API for Hot Wheels collectors to manage their collections,
    trade with other collectors, and explore the complete Hot Wheels catalog.

    ## Features
    - **Authentication**: JWT-based secure authentication system
    - **Catalog**: Browse 50+ years of Hot Wheels models
    - **Collections**: Track and manage personal collections
    - **Marketplace**: Buy, sell, and trade with other collectors
    - **Reviews**: Rate and review transactions

    ## Getting Started
    1. Register a new account via `/api/v1/auth/register`
    2. Login to receive JWT tokens via `/api/v1/auth/login`
    3. Include the access token in the `Authorization` header
    4. Start exploring the catalog and managing your collection!

    ## Rate Limits
    - Anonymous users: 100 requests/minute
    - Authenticated users: 500 requests/minute
    - Rate limit headers included in all responses

    ## Support
    For API support, please contact the development team.

  contact:
    name: API Support
    email: support@hotwheels-collector.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.hotwheels-collector.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/api/v1/auth/login`.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: "Invalid request parameters"
            code:
              type: string
              example: "VALIDATION_ERROR"
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false

    # ==================== User Schemas ====================
...
    Casting:
      type: object
      properties:
        casting_id:
          type: integer
          example: 1
        casting_name:
          type: string
          example: "'67 Camaro"
        first_release_year:
          type: integer
          example: 1968
        manufacturer_id:
          type: integer
          nullable: true
          example: 1
        scale:
          type: string
          example: "1:64"
        designer:
          type: string
          nullable: true
          example: "Harry Bradley"
        based_on_real_car:
          type: boolean
          example: true
        real_car_model:
          type: string
          nullable: true
          example: "1967 Chevrolet Camaro"
        created_at:
          type: string
          format: date-time
    Series:
      type: object
      properties:
        series_id:
          type: integer
          example: 1
        series_name:
          type: string
          example: "Mainline 2024"
        release_year:
          type: integer
          example: 2024
        manufacturer_id:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
          example: "Standard mainline series"
        is_limited_edition:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
    CarModel:
      type: object
      properties:
        model_id:
          type: integer
          example: 1
        casting_id:
          type: integer
          example: 1
        series_id:
          type: integer
          nullable: true
          example: 1
        release_year:
          type: integer
          example: 2024
        color:
          type: string
          example: "Spectraflame Red"
        tampo_design:
          type: string
          nullable: true
          example: "Racing stripes with #7"
        wheel_type:
          type: string
          nullable: true
          example: "5-Spoke"
        base_color:
          type: string
          nullable: true
          example: "Chrome"
        window_color:
          type: string
          nullable: true
          example: "Tinted"
        interior_color:
          type: string
          nullable: true
          example: "Black"
        production_code:
          type: string
          nullable: true
          example: "L2593"
        sku:
          type: string
          nullable: true
          example: "HCW35-N9B1"
        rarity_level:
          type: string
          enum: [Common, Uncommon, Rare, Chase, Super Treasure Hunt]
          example: "Common"
        estimated_production_quantity:
          type: integer
          nullable: true
          example: 50000
        msrp:
          type: number
          format: float
          nullable: true
          example: 1.29
        casting_name:
          type: string
          description: Included when relations are requested
          example: "'67 Camaro"
        series_name:
          type: string
          nullable: true
          description: Included when relations are requested
          example: "Mainline 2024"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Missing required fields: email, password"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Resource not found"
...
paths:
...
  /api/v1/catalog/models:
    get:
      summary: Search car models
      description: |
        Search and filter car models with pagination.

        **Filtering:**
        - Search by casting name with `q` parameter
        - Filter by release year range
        - Filter by rarity level
        - Filter by color
        - Filter by series name
      tags:
        - Catalog
      parameters:
        - name: q
          in: query
          description: Search query (matches casting name)
          schema:
            type: string
          example: "Camaro"
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: Items per page (max 100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: year_min
          in: query
          description: Minimum release year
          schema:
            type: integer
            minimum: 1968
          example: 2020
        - name: year_max
          in: query
          description: Maximum release year
          schema:
            type: integer
          example: 2024
        - name: rarity
          in: query
          description: Rarity level filter
          schema:
            type: string
            enum: [Common, Uncommon, Rare, Chase, Super Treasure Hunt]
        - name: color
          in: query
          description: Color filter
          schema:
            type: string
          example: "Red"
        - name: series
          in: query
          description: Series name filter
          schema:
            type: string
          example: "Mainline"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      models:
                        type: array
                        items:
                          $ref: '#/components/schemas/CarModel'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
  /api/v1/catalog/models/{model_id}:
    get:
      summary: Get a specific car model
      description: Returns detailed information about a specific car model
      tags:
        - Catalog
      parameters:
        - name: model_id
          in: path
          required: true
          description: Car model ID
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Model details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      model:
                        $ref: '#/components/schemas/CarModel'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/catalog/models/rare:
    get:
      summary: Get rare and special models
      description: Returns rare models (Rare, Chase, Super Treasure Hunt)
      tags:
        - Catalog
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Rare models list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      models:
                        type: array
                        items:
                          $ref: '#/components/schemas/CarModel'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
  /api/v1/catalog/models/recent:
    get:
      summary: Get recently released models
      description: Returns models released in the current or specified year
      tags:
        - Catalog
      parameters:
        - name: year
          in: query
          description: Release year (defaults to current year)
          schema:
            type: integer
          example: 2024
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Recent models list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      models:
                        type: array
                        items:
                          $ref: '#/components/schemas/CarModel'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
  /api/v1/catalog/models/popular:
    get:
      summary: Get most popular models
      description: Returns models most commonly found in user collections
      tags:
        - Catalog
      parameters:
        - name: limit
          in: query
          description: Number of results (max 50)
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: Popular models list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      models:
                        type: array
                        items:
                          $ref: '#/components/schemas/CarModel'
  /api/v1/catalog/castings:
    get:
      summary: Get all castings
      description: Returns all castings with pagination
      tags:
        - Catalog
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Castings list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      castings:
                        type: array
                        items:
                          $ref: '#/components/schemas/Casting'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
  /api/v1/catalog/castings/search:
    get:
      summary: Search castings
      description: Search castings by name
      tags:
        - Catalog
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
          example: "Mustang"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      castings:
                        type: array
                        items:
                          $ref: '#/components/schemas/Casting'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/catalog/castings/{casting_id}:
    get:
      summary: Get a specific casting
      description: Returns detailed information about a specific casting
      tags:
        - Catalog
      parameters:
        - name: casting_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Casting details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      casting:
                        $ref: '#/components/schemas/Casting'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/catalog/castings/{casting_id}/models:
    get:
      summary: Get models for a casting
      description: Returns all car model variants for a specific casting
      tags:
        - Catalog
      parameters:
        - name: casting_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Models list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      models:
                        type: array
                        items:
                          $ref: '#/components/schemas/CarModel'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/catalog/series:
    get:
      summary: Get all series
      description: Returns all series with optional year filter
      tags:
        - Catalog
      parameters:
        - name: year
          in: query
          description: Filter by release year
          schema:
            type: integer
          example: 2024
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Series list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      series:
                        type: array
                        items:
                          $ref: '#/components/schemas/Series'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
  /api/v1/catalog/series/{series_id}:
    get:
      summary: Get a specific series
      description: Returns detailed information about a specific series
      tags:
        - Catalog
      parameters:
        - name: series_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Series details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      series:
                        $ref: '#/components/schemas/Series'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/catalog/series/{series_id}/models:
    get:
      summary: Get models in a series
      description: Returns all car models in a specific series
      tags:
        - Catalog
      parameters:
        - name: series_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Models list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      models:
                        type: array
                        items:
                          $ref: '#/components/schemas/CarModel'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        '404':
          $ref: '#/components/responses/NotFound'
...